+++ src/bin/bwx-agent/actions.rs
@@ -6,7 +6,9 @@ mod touchid;
 mod util;
 
 pub use auth::{check_lock, lock, login, register, unlock};
-pub use crypto::{clipboard_store, decrypt, encrypt, version};
+pub use crypto::{
+    clipboard_store, decrypt, decrypt_batch, encrypt, version,
+};
 pub use ssh::{
     decrypt_located_ssh_private_key, get_ssh_public_keys,
     locate_ssh_private_key,